Phase 4.4: extract storage download route
